refactor: [step2] LottoGame에서 LottoShop 분리
